Vanilla vs Neo4j/rag-duo/requirements.txt
chromadb
numpy
tqdm
python-dotenv
requests
neo4j
pydantic
fastapi[standard]
uvicorn
python-dateutil
sentence-transformers
